+++ scripts/twenty-one/index.js
@@ -12,3 +12,4 @@ export * from "./phases/core.js";
 export * from "./phases/turn.js";
 export * from "./phases/special.js";
 export * from "./rulesets/index.js";
+export * from "./rulesets/goblin.js";